refactor(patchers): reorganize folder structure into OriginalEngines and Trx
Splits the flat Gameflows/ namespace into:
- OriginalEngines/Models/ — raw data models (Gameflow, Sequence, TpcStringArray, etc.)
- OriginalEngines/Parsers/ — GameflowReader and SequenceOpcodeExtensions
- OriginalEngines/ — shared Constants
- Trx/Patchers/ — TrxGameflowPatcher (was Trx/)

Updates all namespaces and using directives accordingly.
