+++ src/TombLauncher.Patchers/OriginalEngines/Models/GameflowLanguage.cs
@@ -1,4 +1,4 @@
-namespace TombLauncher.Patchers.Gameflows;
+namespace TombLauncher.Patchers.OriginalEngines.Models;
 
 public enum GameflowLanguage
 {
