+++ src/TombLauncher.Patchers/OriginalEngines/Models/SequenceOpcode.cs
@@ -1,4 +1,4 @@
-namespace TombLauncher.Patchers.Gameflows;
+namespace TombLauncher.Patchers.OriginalEngines.Models;
 
 public enum SequenceOpcode
 {
