+++ src/TombLauncher.Patchers/Trx/Patchers/TrxGameflowPatcher.cs
@@ -1,0 +1,6 @@
+namespace TombLauncher.Patchers.Trx.Patchers;
+
+public class TrxGameflowPatcher
+{
+    
+}
