--- src/TombLauncher.Patchers/Trx/TrxGameflowPatcher.cs
@@ -1,6 +1,0 @@
-namespace TombLauncher.Patchers.Trx;
-
-public class TrxGameflowPatcher
-{
-    
-}
